+++ ios/Classes/core/PMManager.m
@@ -69,7 +69,9 @@
                 if (collection.assetCollectionSubtype == PHAssetCollectionSubtypeSmartAlbumUserLibrary) {
                     PMAssetPathEntity *pathEntity = [PMAssetPathEntity
                         entityWithId:collection.localIdentifier
-                                name:collection.localizedTitle];
+                                name:collection.localizedTitle
+                     assetCollection:collection
+                    ];
                     pathEntity.isAll = YES;
                     [array addObject:pathEntity];
                     break;
@@ -78,24 +80,31 @@
         }
         return array;
     }
-    [self logCollections:smartAlbumResult option:assetOptions];
-    [self injectAssetPathIntoArray:array
-                            result:smartAlbumResult
-                           options:assetOptions
-                            hasAll:hasAll
-                  containsModified:option.containsModified];
 
-    PHFetchResult<PHAssetCollection *> *albumResult = [PHAssetCollection
-        fetchAssetCollectionsWithType:PHAssetCollectionTypeAlbum
-                              subtype:PHAssetCollectionSubtypeAny
-                              options:fetchCollectionOptions];
-    [self logCollections:albumResult option:assetOptions];
-    [self injectAssetPathIntoArray:array
-                            result:albumResult
-                           options:assetOptions
-                            hasAll:hasAll
-                  containsModified:option.containsModified];
+    if ([pathFilterOption.type indexOfObject:@(PHAssetCollectionTypeSmartAlbum)] != NSNotFound) {
+        [self logCollections:smartAlbumResult option:assetOptions];
+        [self injectAssetPathIntoArray:array
+                                result:smartAlbumResult
+                               options:assetOptions
+                                hasAll:hasAll
+                      containsModified:option.containsModified
+                      pathFilterOption:pathFilterOption
+        ];
+    }
 
+    if ([pathFilterOption.type indexOfObject:@(PHAssetCollectionTypeAlbum)] != NSNotFound) {
+        PHFetchResult<PHAssetCollection *> *albumResult = [PHAssetCollection
+            fetchAssetCollectionsWithType:PHAssetCollectionTypeAlbum
+                                  subtype:PHAssetCollectionSubtypeAny
+                                  options:fetchCollectionOptions];
+        [self logCollections:albumResult option:assetOptions];
+        [self injectAssetPathIntoArray:array
+                                result:albumResult
+                               options:assetOptions
+                                hasAll:hasAll
+                      containsModified:option.containsModified
+                      pathFilterOption:pathFilterOption];
+    }
     return array;
 }
 
@@ -192,7 +201,8 @@
                           result:(PHFetchResult *)result
                          options:(PHFetchOptions *)options
                           hasAll:(BOOL)hasAll
-                containsModified:(BOOL)containsModified {
+                containsModified:(BOOL)containsModified
+                pathFilterOption:(PMPathFilterOption *)pathFilterOption {
     for (id collection in result) {
         if (![collection isKindOfClass:[PHAssetCollection class]]) {
             continue;
@@ -200,11 +210,11 @@
 
         PHAssetCollection *assetCollection = (PHAssetCollection *) collection;
 
-        // Check whether it's "Recently Deleted"
-        if (assetCollection.assetCollectionSubtype == PHAssetCollectionSubtypeSmartAlbumRecentlyAdded
-            || assetCollection.assetCollectionSubtype == 1000000201) {
-            continue;
-        }
+//        // Check whether it's "Recently Deleted"
+//        if (assetCollection.assetCollectionSubtype == PHAssetCollectionSubtypeSmartAlbumRecentlyAdded
+//            || assetCollection.assetCollectionSubtype == 1000000201) {
+//            continue;
+//        }
 
         // Check nullable id and name
         NSString *localIdentifier = assetCollection.localIdentifier;
@@ -213,7 +223,9 @@
             continue;
         }
 
-        PMAssetPathEntity *entity = [PMAssetPathEntity entityWithId:localIdentifier name:localizedTitle];
+//        [[PMLogUtils sharedInstance] debug:[NSString stringWithFormat:@"id: %@, title: %@, type: %d subType: %d", localIdentifier, localizedTitle, (int)assetCollection.assetCollectionType, (int)assetCollection.assetCollectionSubtype]];
+
+        PMAssetPathEntity *entity = [PMAssetPathEntity entityWithId:localIdentifier name:localizedTitle assetCollection:assetCollection];
         entity.isAll = assetCollection.assetCollectionSubtype == PHAssetCollectionSubtypeSmartAlbumUserLibrary;
 
         if (containsModified) {
@@ -225,10 +237,15 @@
             }
         }
 
-        if (!hasAll && entity.isAll) {
+        if (hasAll && entity.isAll) {
+            [array addObject:entity];
             continue;
         }
-        [array addObject:entity];
+
+        if ([pathFilterOption.subType indexOfObject:@(PHAssetCollectionSubtypeAny)] != NSNotFound ||
+            [pathFilterOption.subType indexOfObject:@(assetCollection.assetCollectionSubtype)] != NSNotFound) {
+            [array addObject:entity];
+        }
     }
 }
 
@@ -851,11 +868,11 @@
     ![info[PHImageResultIsDegradedKey] boolValue]; // thumbnail
 }
 
-- (PMAssetPathEntity *)fetchPathProperties:(NSString *)id type:(int)type filterOption:(NSObject<PMBaseFilter> *)filterOption {
+- (PMAssetPathEntity *)fetchPathProperties:(NSString *)id type:(int)type filterOption:(NSObject <PMBaseFilter> *)filterOption {
     PHFetchOptions *collectionFetchOptions = [PHFetchOptions new];
     PHFetchResult<PHAssetCollection *> *result = [PHAssetCollection
-                                                  fetchAssetCollectionsWithLocalIdentifiers:@[id]
-                                                  options:collectionFetchOptions];
+        fetchAssetCollectionsWithLocalIdentifiers:@[id]
+                                          options:collectionFetchOptions];
 
     if (result == nil || result.count == 0) {
         return nil;
@@ -869,7 +886,9 @@
         return nil;
     }
     PMAssetPathEntity *entity = [PMAssetPathEntity entityWithId:localIdentifier
-                                                           name:localizedTitle];
+                                                           name:localizedTitle
+                                                assetCollection:collection
+    ];
     entity.isAll = collection.assetCollectionSubtype == PHAssetCollectionSubtypeSmartAlbumUserLibrary;
     return entity;
 }