+++ lib/photo_manager.dart
@@ -8,6 +8,7 @@
 library photo_manager;
 
 export 'src/filter/base_filter.dart';
+export 'src/filter/path_filter.dart';
 export 'src/filter/classical/filter_option_group.dart';
 export 'src/filter/classical/filter_options.dart';